// client/dist/widgets/common/search/src/config.ts
export const DEFAULT_GEOCODE_KEY = 'default_geocode'

export const MAX_SUGGESTION = 6
export const MAX_RECENT_SEARCHES = 6
export const DEFAULT_SPATIAL_REFERENCE = {
  wkid: 4326
}

export enum SearchServiceType {
  GeocodeService = 'GeocodeService',
  FeatureService = 'FeatureService'
}

export interface InitResultServiceListOption {
  configId?: string
  magicKey?: string
  isFromSuggestion?: boolean
}

export enum SearchResultView {
  ResultPanel = 'Result_Panel',
  OtherWidgets = 'Other_Widgets'
}

export enum ArrangementStyle {
  Style1 = 'Style1',
  Style2 = 'Style2',
  Style3 = 'Style3'
}

export enum SearchResultStyle {
  Classic = 'Classic',
  Compact = 'Compact'
}

export enum EntityStatusType {
  None = '',
  Init = 'init',
  Loading = 'loading',
  Loaded = 'loaded',
  Warning = 'warning',
  Error = 'error',
}

export enum SourceType {
  CustomSearchSources = 'CustomSearchSources',
  MapCentric = 'MapCentric',
}

export interface SearchStatus {
  serviceEnabledList?: string[]
  searchText?: string
  status?: InitResultServiceListOption
}

export interface SearchFieldData {
  [key: string]: {
    searchFields: FieldSchema[]
    displayFields: FieldSchema[]
  }
}

export interface DatasourceCreatedStatus {
  [key: string]: DataSourceStatus
}

export type IMDatasourceCreatedStatus = ImmutableObject<DatasourceCreatedStatus>

export interface NewLayerConfigItemSetting extends SearchLayerDataConfig {
  enable: boolean
}

export interface NewGeocodeItemSetting extends SearchGeocodeDataConfig {
  enable: boolean
}

export interface NewDatasourceConfigItem extends SearchDataConfig {
  enable: boolean
  geocodeURL?: string
}
// ...
export interface SearchResult {
  [key: string]: string[]
}

export type IMRecordResultType = ImmutableObject<RecordResultType>
export type IMSearchResult = ImmutableObject<SearchResult>

export interface SelectionList {
  [key: string]: ImmutableArray<string>
}

export type IMSelectionList = ImmutableObject<SelectionList>

export interface SuggestionItem {
  suggestionHtml: string | Element
  suggestion: string
  isRecentSearch?: boolean
  configId?: string
  isFromSuggestion?: boolean
  magicKey?: string
}

export interface Suggestion {
  suggestionItem: SuggestionItem[]
  layer: string
  icon?: any
  err?: any
}

export type IMSuggestion = ImmutableObject<Suggestion>

export interface DatasourceList {
  [key: string]: DatasourceListItem
}

export interface GeocodeList {
  [key: string]: GeocodeListItem
}

export type ServiceList = DatasourceList & GeocodeList
// ...
export type IMServiceList = ImmutableObject<ServiceList>
export type IMDatasourceListItem = ImmutableObject<DatasourceListItem>

export type IMDataSourceConfigWithMapCentric = ImmutableObject<DataSourceConfigWithMapCentric>

export interface config extends SearchSuggestionConfig {
  synchronizeSettings?: boolean
  sourceType?: SourceType
  dataSourceConfigWithMapCentric?: DataSourceConfigWithMapCentric
  datasourceConfig?: SearchDataConfig[]
  hint: string
  searchResultView: SearchResultView
  arrangementStyle: ArrangementStyle
  searchResultStyle: SearchResultStyle
  resultMaxNumber: number
  isAutoSelectFirstResult: boolean
  // link
  linkParam?: IMLinkParam
}

export type IMConfig = ImmutableObject<config>